requirements.txt
python-telegram-bot>=21.0,<22.0
google-genai>=1.0.0
